requirements.txt
streamlit
Pillow
ollama
